Fixed a compile error with int vs long unsigned int, in the gfx lab.
git-svn-id: https://wiki.csc.calpoly.edu/svn/AsteroidBlaster@734 342171fe-4ba3-422f-b357-09b6f24955fe

+++ Weapons/TractorBeam.h
@@ -23,7 +23,7 @@ class TractorBeam : public Weapon {
       virtual void debug();
       virtual void fire();
    private:
-      int lastFiredFrame;
+      long unsigned int lastFiredFrame;
       bool soundPlaying;
       int soundHandle;
 };
